+++ ingest2/generator/id_generator.go
@@ -44,3 +44,19 @@ func (g *ID) Next() int64 {
 	g.seq++
 	return (int64(g.Significant) << 32) | int64(g.seq)
 }
+
+func MakeIDUint32(significant int32, sequence uint32) int64 {
+	if significant < 0 {
+		panic(errors.New("negative significant is not supported"))
+	}
+
+	return (int64(significant) << 32) | int64(sequence)
+}
+
+func MakeIDUint16(significant int32, seq1, seq2 uint16) int64 {
+	if significant < 0 {
+		panic(errors.New("negative significant is not supported"))
+	}
+
+	return (int64(significant) << 32) | int64(seq1)<<16 | int64(seq2)
+}